refactor(pruner): extract test helpers into helpers_test.go
